refactor: internal module for Api
